Just use ints to store reserved mp. Technically this introduces a bunch of breakpoints, but the implementation is cleaner and the interface doesn't sometimes lie about the mp you lose. Very minor global buff to casting a lot of permabuffs, also allows you to reserve down to 0 mp remaining rather than forcing you to keep 1 mp unreserved.

+++ crawl-ref/source/spl-cast.h
@@ -104,7 +104,7 @@ vector<string> desc_success_chance(const monster_info& mi, int pow, bool evoked,
 spret_type your_spells(spell_type spell, int powc = 0, bool allow_fail = true,
                        const item_def* const evoked_item = nullptr);				   
 					   
-double spell_mp_freeze(spell_type spell);
+int spell_mp_freeze(spell_type spell);
 bool is_buff_spell (spell_type spell);
 extern const char *fail_severity_adjs[];
